requirements.txt
anyio==3.5.0
apispec==5.2.1
apispec-webframeworks==0.5.2
appdirs==1.4.4
asgiref==3.5.1
asttokens==2.0.5
audioread==2.1.9
backcall==0.2.0
gunicorn==19.9.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
elastic-transport==8.1.2
elasticsearch==7.13.2
executing==0.8.3
fastapi==0.76.0
Flask==2.1.2
flask-swagger-ui==3.36.0
fonttools==4.33.3
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-speech==2.14.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.13.0
idna==3.3
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.71
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.38.1
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
numba==0.55.2
numpy==1.22.4
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pipwin==0.5.2
pooch==1.6.0
prompt-toolkit==3.0.29
proto-plus==1.20.5
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydub==0.25.1
Pygments==2.12.0
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.8
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-multipart==0.0.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
stack-data==0.2.0
starlette==0.18.0
threadpoolctl==3.1.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
Werkzeug==2.1.2
